Reorganize html into folder
