+++ src/html/add-book.html
@@ -2,8 +2,8 @@
 <html>
     <head>
         <meta charset="UTF-8" />
-        <link rel="stylesheet" href=".././dist/output.css" />
-        <title>Papyrus @ Athena</title>
+        <link rel="stylesheet" href="../../dist/output.css" />
+        <title>Add New Book - Papyrus @ Athena</title>
     </head>
     <body class="p-8">
         <h1
@@ -104,6 +104,6 @@
             >
         </footer>
 
-        <script type="module" src="./js/add-book.js"></script>
+        <script type="module" src="../js/add-book.js"></script>
     </body>
 </html>
