+++ src/html/index-old.html
@@ -2,7 +2,7 @@
 <html>
     <head>
         <meta charset="UTF-8" />
-        <link rel="stylesheet" href=".././dist/output.css" />
+        <link rel="stylesheet" href="../../dist/output.css" />
         <title>Papyrus @ Athena</title>
     </head>
     <body>
