+++ src/html/index.html
@@ -2,7 +2,7 @@
 <html>
     <head>
         <meta charset="UTF-8" />
-        <link rel="stylesheet" href=".././dist/output.css" />
+        <link rel="stylesheet" href="../../dist/output.css" />
         <title>Papyrus @ Athena</title>
     </head>
     <body class="p-8">
@@ -54,9 +54,14 @@
                 href="./add-book.html"
                 class="text-blue-700 font-bold text-xl px-5"
                 >Add Books</a
+            >|
+            <a
+                href="./reports/reports.html"
+                class="text-blue-700 font-bold text-xl px-5"
+                >Reports</a
             >
         </footer>
 
-        <script type="module" src="./js/index.js"></script>
+        <script type="module" src="../js/index.js"></script>
     </body>
 </html>
