+++ src/html/member-master.html
@@ -2,8 +2,8 @@
 <html>
     <head>
         <meta charset="UTF-8" />
-        <link rel="stylesheet" href=".././dist/output.css" />
-        <title>Papyrus @ Athena</title>
+        <link rel="stylesheet" href="../../dist/output.css" />
+        <title>Members - Papyrus @ Athena</title>
     </head>
     <body class="p-8">
         <h1
@@ -16,7 +16,7 @@
             <div
                 id="message-container"
                 class="text-center w-full mt-8 text-2xl"
-            ></div>
+            >Please enter flat number and click the search button to check if member already exists. If no member is found you can add the name and save.</div>
             <form
                 id="member-search-form"
                 action=""
@@ -77,6 +77,6 @@
             >
         </footer>
 
-        <script type="module" src="./js/member-master.js"></script>
+        <script type="module" src="../js/member-master.js"></script>
     </body>
 </html>
