+++ src/html/reports/reports.html
@@ -1,0 +1,32 @@
+<!DOCTYPE html>
+<html>
+    <head>
+        <meta charset="UTF-8" />
+        <link rel="stylesheet" href="../../../dist/output.css" />
+        <title>Library Reports - Papyrus @ Athena</title>
+    </head>
+    <body class="p-8">
+        <h1
+            class="text-4xl text-white w-11/12 m-auto bg-gray-700 text-center p-4 rounded-lg bg-opacity-50 font-extrabold mb-10"
+        >
+            Papyrus @ Athena
+        </h1>
+        <main class="w-11/12 m-auto">
+            <h1>Reports</h1>
+            <ul>
+                <li>Outstanding books</li>
+            </ul>
+        </main>
+
+        <footer class="flex justify-center fixed bottom-5 w-full">
+            <a
+                href="./index.html"
+                class="text-blue-700 font-bold text-xl px-5"
+            >
+                Home </a
+            >|
+        </footer>
+
+        <script type="module" src="../../js/index.js"></script>
+    </body>
+</html>
